requirements.txt
certifi==2018.8.24
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
decorator==4.0.11
dj-database-url==0.5.0
Django==2.0.6
django-cors-headers==2.2.0
django-heroku==0.3.1
django-rest-swagger==2.1.2
django-summernote==0.8.8.6
django-templated-mail==1.1.1
djangorestframework==3.7.7
djangorestframework-jsonapi==2.4.0
djangorestframework-jwt==1.11.0
djoser==1.1.5
gunicorn==19.9.0
idna==2.7
imageio==2.1.2
inflection==0.3.1
InstagramAPI==1.0.2
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
moviepy==0.2.3.2
numpy==1.15.1
openapi-codec==1.3.2
Pillow==5.2.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
PyJWT==1.6.4
python-dateutil==2.7.3
python-decouple==3.1
pytz==2018.5
requests==2.11.1
requests-toolbelt==0.7.0
simplejson==3.16.0
six==1.11.0
tqdm==4.11.2
uritemplate==3.0.0
urllib3==1.23
whitenoise==4.1
